Fixes STRIPES-33 ("Run the Okapi Console from git checkout without uncommitted changes to webpack.config.cli.js")
devServer.js now consults a file called stripes-modules.js for
additional Stripes modules to include as well as those specified by
webpack.config.cli.js. If the file is present, then its contents are
added to the module-list; if not, its absense is unobtrusively noted
and all proceeds as before. The stripes-modules.js file itself is not
kept in git, since different developers will want it to have different
contents at different times; but a stripes-modules.js.example file is
provided to show the format.